+++ deployment/terraform/resources/rg.tf
@@ -1,4 +1,8 @@
 resource "azurerm_resource_group" "pc" {
   name     = "${local.prefix}_rg"
   location = var.region
+
+  tags = {
+    "ringValue" = "r0"
+  }
 }