+++ pccommon/pccommon/cli.py
@@ -31,7 +31,14 @@ def load(
             account_url=account_url, sas_token=sas, table_name=table
         )
         for coll_id, config in rows.items():
-            col_config_table.set_config(coll_id, CollectionConfig(**config))
+            print("Loading config for collection", coll_id)
+            try:
+                col_config_table.set_config(coll_id, CollectionConfig(**config))
+            except Exception as e:
+                print("========================================")
+                print(f"Error loading config for collection {coll_id}: {e}")
+                print(f"{coll_id} has been skipped!")
+                print("========================================")
 
     elif type == "container":
         cont_config_table = ContainerConfigTable.from_sas_token(
@@ -66,6 +73,7 @@ def dump(sas: str, account: str, table: str, type: str, **kwargs: Any) -> int:
         else:
             for (_, collection_id, col_config) in col_config_table.get_all():
                 assert collection_id
+                assert col_config
                 result[collection_id] = col_config.dict()
 
     elif type == "container":
@@ -80,6 +88,7 @@ def dump(sas: str, account: str, table: str, type: str, **kwargs: Any) -> int:
             result[f"{con_account}/{id}"] = con_config.dict()
         else:
             for (storage_account, container, con_config) in con_config_table.get_all():
+                assert con_config
                 result[f"{storage_account}/{container}"] = con_config.dict()
     else:
         print(f"Unknown type: {type}")